Reorganize the carpaccio client targets (library, tests, executable).
