+++ clients/cpp/Client/executable/sources/main.cpp
@@ -1,0 +1,12 @@
+
+#include <extreme_carpaccio/client/Client.hpp>
+
+#include <cstdlib>
+
+int main()
+{
+   extreme_carpaccio::client::CarpaccioServer server;
+   server.start();
+
+   return EXIT_SUCCESS;
+}
