+++ clients/cpp/Client/library/sources/extreme_carpaccio/client/HttpConfig.hpp
@@ -1,7 +1,14 @@
+
 #ifndef EXTREME_CARPACCIO_CLIENT_HTTPCONFIG_HPP
 #define EXTREME_CARPACCIO_CLIENT_HTTPCONFIG_HPP
 
 #define HTTP_SERVER_IP "127.0.0.1"
 #define HTTP_SERVER_PORT 8081
+
+namespace extreme_carpaccio {
+namespace client {
+
+} // namespace client
+} // namespace extreme_carpaccio
 
 #endif // EXTREME_CARPACCIO_CLIENT_HTTPCONFIG_HPP
