--- clients/cpp/src/main.cpp
@@ -1,8 +1,0 @@
-
-#include <extreme_carpaccio_client/ExtremeCarpaccioClient.hpp>
-
-int main()
-{
-   extreme_carpaccio_client::CarpaccioServer server;
-   server.start();
-}
